requirements.txt

python-telegram-bot==20.3
motor==3.1.1
pymongo==4.3.3
python-dotenv==1.0.0
APScheduler==3.10.1
